% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_by_column.R
\name{abbreviate_by_column}
\alias{abbreviate_by_column}
\title{Replace Column Values with Shorter Version}
\usage{
abbreviate_by_column(df, col_names = NULL, remove_vowels = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{col_names}{A character vector specifying the names of the columns to be abbreviated.
If NULL (default), the function attempts to use the first character column.}

\item{remove_vowels}{A logical value indicating whether to remove vowels from column values after the first character. Default is FALSE.}
}
\value{
A data frame with specified columns abbreviated.
}
\description{
This function shortens the character column values in a data frame by removing vowels after the first character,
and also provides an option to disable this vowel removal. It replaces spaces with underscores and removes characters after a slash.
If no column names are provided, the function attempts to find and use the first character column in the data frame.
}
\examples{
df <- data.frame(
  col1 = c("Hutan lahan kering sekunder / bekas tebangan", "Savanna / Padang rumput"),
  col2 = c("Hutan lahan kering sekunder", "Savanna"),
  stringsAsFactors = FALSE
)
abbreviate_by_column(df, c("col1", "col2"), remove_vowels=TRUE)
}
