% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lookup_table.R
\name{apply_lookup_table}
\alias{apply_lookup_table}
\title{Apply legend and color table to raster files}
\usage{
apply_lookup_table(raster_file, lookup_lc)
}
\arguments{
\item{raster_file}{A raster file}

\item{lookup_lc}{A lookup table that maps raster values to legend and colors}
}
\value{
A raster file with the specified  legend and color table
}
\description{
Apply legend and color table to raster files
}
\examples{

# Load the LUMENSR package
library(LUMENSR)
library(terra)

# Load the 2011 land cover raster for the Kalbar region
kalbar_2011 <- rast(LUMENSR_example("kalbar_LC11.tif")) \%>\% classify(cbind(0,NA))

# apply a lookup table into a table
kalbar_2011_legend <- apply_lookup_table(kalbar_2011, lc_lookup_klhk)

# Plot the classified raster
plot(kalbar_2011_legend)
}
\author{
Dony Indiarto
}
