% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_time_period.R
\name{assign_time_period}
\alias{assign_time_period}
\title{Assign a time period to a SpatRaster object}
\usage{
assign_time_period(raster_layer, year_)
}
\arguments{
\item{raster_layer}{A SpatRaster object}

\item{year_}{A string representing a year}
}
\value{
The original SpatRaster object, with an assigned time period
}
\description{
Assign a time period to a SpatRaster object
}
\examples{
# Load terra package
library(terra)
# Create example SpatRaster
r <- rast(nrow=10, ncol=10)
r[] <- 1:ncell(r)
# Assign a time period
assign_time_period(r, "2023")
}
