% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conv_factor_to_ha.R
\name{calc_res_conv_factor_to_ha}
\alias{calc_res_conv_factor_to_ha}
\title{Calculate Resolution Conversion Factor To Hectares}
\usage{
calc_res_conv_factor_to_ha(raster_input)
}
\arguments{
\item{raster_input}{A terra::SpatRaster object.}
}
\value{
A numerical value representing the conversion factor of the raster map resolution to hectares.
}
\description{
This function calculates the conversion factor of a raster map resolution to hectares,
depending on the coordinate reference system (CRS) of the raster.
Raster maps with projection in meter or degree units are supported.
}
