% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_top_lcc.R
\name{calc_top_lcc}
\alias{calc_top_lcc}
\title{Get the top n rows of the input data frame}
\usage{
calc_top_lcc(crosstab_result, n_rows)
}
\arguments{
\item{crosstab_result}{A data frame of  a frequency table crosstab_result must have a column named "Freq", representing the number of pixels / area for a given land cover change
(crosstab) of two or more land cover maps,}

\item{n_rows}{The number of rows to select.}
}
\value{
A data frame containing the top n rows (top n changes.
}
\description{
This function filters rows that have more than one distinct value
across all columns except the last one, and then selects the top n rows
based on the value of the last column.
}
