% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trajectory_map.R
\name{calc_trajectory_map}
\alias{calc_trajectory_map}
\title{Calculate Trajectory Map}
\usage{
calc_trajectory_map(
  lc_t1_,
  lc_t2_,
  lookup_traj_reclass,
  lookup_trajectory_complete,
  trajectory_column_name = "trajectory"
)
}
\arguments{
\item{lc_t1_}{A raster object representing the land cover at time 1.}

\item{lc_t2_}{A raster object representing the land cover at time 2.}

\item{lookup_traj_reclass}{A data frame containing the reclassification rules.}

\item{lookup_trajectory_complete}{A data frame containing the lookup table for trajectories.}

\item{trajectory_column_name}{A string representing the column name for the trajectory in the lookup table.}
}
\value{
A raster object representing the trajectory map.
}
\description{
This function calculates the trajectory map by reclassifying two rasters
(representing land cover at two different time points), concatenating them,
and then adding categorical information based on a provided lookup table.
}
\examples{
\dontrun{
# assuming we have two raster objects "lc_t1", "lc_t2",
#a reclassification table "lookup_traj_reclass",
# and a lookup table "lookup_trajectory_complete"
my_trajectory_map <- calc_trajectory_map(lc_t1_, lc_t2_, lookup_traj_reclass,
lookup_trajectory_complete,
trajectory_column_name = "trajectory")
}

}
