% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_lookup.R
\name{color_forest_trajectories}
\alias{color_forest_trajectories}
\title{Create a Color Lookup Table for Forest Definitions}
\usage{
color_forest_trajectories()
}
\value{
A tibble containing forest definitions and their corresponding color codes
}
\description{
This function generates a tibble that maps forest definitions to their corresponding color codes.
The function will return a tibble containing two columns: 'def' for forest definitions and 'hex_color' for color codes.
}
