% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_lookup.R
\name{color_landuse_trajectories}
\alias{color_landuse_trajectories}
\title{Create a Color Lookup Table for Trajectory Categories}
\usage{
color_landuse_trajectories()
}
\value{
A tibble containing trajectory categories and their corresponding color codes
}
\description{
This function generates a tibble that maps trajectory categories to their corresponding color codes.
The function returns a tibble containing two columns: 'trajectory' for the categories and 'hex_color' for the color codes.
}
