% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crosstab.R
\name{create_crosstab}
\alias{create_crosstab}
\title{Create Crosstab From Raster List}
\usage{
create_crosstab(land_cover, zone)
}
\arguments{
\item{land_cover}{A list of 'SpatRaster' objects.}

\item{zone}{A SpatRaster' object of planning unit/zone}
}
\value{
A data frame representing the crosstab of the input rasters.
}
\description{
This function takes a list of raster files, creates a stack, and returns a frequency table
(crosstab) of the stack. The frequency table is filtered to remove rows where the frequency is zero.
}
\examples{
\dontrun{
  # Read in the legend for raster data
  subset_legend <- LUMENSR::lc_lookup_klhk

  # Read raster files
  raster_files <- c("kalbar_LC11.tif", "kalbar_LC15.tif", "kalbar_LC20.tif") \%>\%
    map(LUMENSR_example) \%>\%
    map(rast)

  # Loop through raster files
  land_cover <- map(raster_files,
                            ~apply_lookup_table(raster_file = .x,
                                                lookup_lc = subset_legend))

  # Turn raster files into a frequency table
  crosstab_result <- create_crosstab(land_cover)
}

}
