% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_luc_trajectories.R
\name{create_lut_reclass}
\alias{create_lut_reclass}
\title{Create lookup table for land cover trajectories}
\usage{
create_lut_reclass()
}
\value{
A tibble with two columns: \code{ID} (the ID of each land cover
classification) and \code{class} (the name of each land cover class).
}
\description{
This function generates a tibble that maps reclassify IDs to their
corresponding land cover new class names.
}
\examples{
lut_class <- create_lut_reclass()
head(lut_reclass)
}
