% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sankey.R
\name{create_sankey}
\alias{create_sankey}
\title{Create a Sankey Diagram}
\usage{
create_sankey(
  crosstab_,
  max_categories = 15,
  font_size = 10,
  sinks_right = TRUE,
  label_show_varname = FALSE,
  node_position_automatic = TRUE,
  hovertext_show_percentages = FALSE,
  label_show_percentages = FALSE,
  label_show_counts = FALSE,
  link_color = "Source",
  variables_share_values = TRUE
)
}
\arguments{
\item{crosstab_}{A frequency table to be processed.}

\item{max_categories}{The maximum number of categories. Default is 15.}

\item{font_size}{The font size. Default is 10.}

\item{sinks_right}{If TRUE, sinks are on the right. Default is TRUE.}

\item{label_show_varname}{If FALSE, variable names are not shown. Default is FALSE.}

\item{node_position_automatic}{If TRUE, node position is automatic. Default is TRUE.}

\item{hovertext_show_percentages}{If FALSE, hover text does not show percentages. Default is FALSE.}

\item{label_show_percentages}{If FALSE, labels do not show percentages. Default is FALSE.}

\item{label_show_counts}{If FALSE, labels do not show counts. Default is FALSE.}

\item{link_color}{The color of the link. Default is "Source".}

\item{variables_share_values}{If TRUE, variables share values. Default is TRUE.}
}
\value{
A Sankey diagram.
}
\description{
This function takes a frequency table (crosstab) as input, filters the table to only include
rows where the frequency is greater than the average frequency, and creates a Sankey diagram
from the filtered table.
}
\examples{
\dontrun{
  # Assume you have a crosstab_
  sankey_diagram <- create_sankey(crosstab_)
}

}
