% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{driluc_traj_lookup}
\alias{driluc_traj_lookup}
\title{Lookup Table for Major Land Use Types}
\format{
A tibble with 8 rows and 2 columns:
\describe{
\item{traj_id}{Trajectory ID for each major land use type, ranging from 1 to 8}
\item{trajectory}{English descriptions corresponding to each trajectory ID, representing different land use types}
}
}
\source{
The table may be part of a larger data set or analysis within your organization or project. Include additional source information as needed.
}
\usage{
driluc_traj_lookup
}
\description{
This lookup table represents major land use types that are part of the land use change
trajectory analysis. It is used to reclassify a land cover map, mapping specific trajectory IDs
to their respective descriptions.
}
\keyword{datasets}
