% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_luc_trajectories.R
\name{generate_combinations}
\alias{generate_combinations}
\title{Generate Combinations of Trajectory IDs}
\usage{
generate_combinations(df)
}
\arguments{
\item{df}{A data frame containing trajectory IDs and classes.}
}
\value{
A data frame with combinations of trajectory IDs and classes.
}
\description{
This function generates combinations of trajectory IDs from a lookup table and creates a new ID column.
}
\examples{
\dontrun{
lut_class <- create_lut_reclass()
generate_combinations(new_traj_class)
}
}
