% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_to_image.R
\name{ggplot_to_image}
\alias{ggplot_to_image}
\title{Save ggplot2 plot as an Image Object}
\usage{
ggplot_to_image(
  plot_object,
  image_width = 10,
  image_height = 10,
  image_resolution = 300
)
}
\arguments{
\item{plot_object}{ggplot2 plot that needs to be saved as image.}

\item{image_width}{Width of the image in inches. Default is 10.}

\item{image_height}{Height of the image in inches. Default is 10.}

\item{image_resolution}{Resolution of the image in dpi. Default is 300.}
}
\value{
A magick image object of the ggplot2 plot.
}
\description{
This function saves a ggplot2 plot as a magick image object, using the desired width and height in inches and dpi resolution.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mpg, aes(displ, hwy, colour = class)) + geom_point()
img <- ggsave_to_image(p)
}
}
