% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_raster.R
\name{harmonise_raster}
\alias{harmonise_raster}
\title{Crop and harmonise predictors, such as climate and soil data}
\usage{
harmonise_raster(dat_clim, batas_kab)
}
\arguments{
\item{dat_clim}{a spatRaster data}

\item{batas_kab}{the area of interest represented as an sf polygon}
}
\value{
a Spat raster object of the predictor with the same extent and area of batas_kab
A SpatRaster object representing the cropped and harmonized predictor, with the same extent and area as the \code{batas_kab} object.
}
\description{
This function takes a raster data and a polygon of the area of interest,
crops the raster data to the area of interest, and harmonizes the projection
and extent of the cropped raster data to match that of the polygon.
}
\examples{
worldclim_temp_indo <- LUMENSR_example("wc2.1_30s_bio_1_indo.tif")
temp_indo <-  terra::rast(worldclim_temp_indo)
harmonise_raster(dat_clim = temp_indo, batas_kab = ntt_admin)
}
