% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lc_lookup_klhk}
\alias{lc_lookup_klhk}
\title{Land cover data of West Kalimantan of 2011}
\format{
A RasterLayer object of the land cover map.

A RasterLayer object of the land cover map.

A RasterLayer object of the land cover map.

A data frame with 23 rows and 2 columns:
\describe{
\item{Value}{The land cover code}
\item{PL20}{The land cover description}
}
}
\source{
Ministry of Forestry and Environment of Indonesia
Land cover data of West Kalimantan of 2015

Ministry of Forestry and Environment of Indonesia
Land cover data of West Kalimantan of 2020

Ministry of Forestry and Environment of Indonesia
The Indonesian Ministry of Environment and Forestry (KLHK) Land Cover Lookup Table

The data was obtained from the Indonesian Ministry of Environment and Forestry (KLHK).
}
\usage{
lc_lookup_klhk
}
\description{
KLHK land cover map of West Kalimantan province for the year 2011.

KLHK land cover map of West Kalimantan province for the year 2015.

KLHK land cover map of West Kalimantan province for the year 2020.

A lookup table that maps land cover codes to land cover descriptions for the Indonesian Ministry of Environment and Forestry (KLHK).
}
\examples{
data(lc_lookup_klhk)
head(lc_lookup_klhk)
}
\keyword{datasets}
