% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lc_lookup_klhk_sequence}
\alias{lc_lookup_klhk_sequence}
\title{Reclassified Land Cover Lookup Table from the Indonesian Ministry of Environment and Forestry (KLHK)}
\format{
A data frame with 23 rows and 2 columns:
\describe{
\item{Value}{Reclassified land cover codes ranging from 1 to 23}
\item{PL20}{Descriptions for each corresponding land cover code}
\item{color_palette}{hex color codes of predefined colour for each land cover class}
}
}
\source{
The data was obtained and adapted from the Indonesian Ministry of Environment and Forestry (KLHK).
}
\usage{
lc_lookup_klhk_sequence
}
\description{
This lookup table relates reclassified land cover codes to their respective descriptions,
as defined by the Indonesian Ministry of Environment and Forestry (KLHK). The original land cover
values have been replaced with a sequence of numbers from 1 to 23.
}
\examples{
data(lc_lookup_klhk_sequence)
head(lc_lookup_klhk_sequence)
}
\keyword{datasets}
