% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcc_summary_by_pu.R
\name{lcc_summary_by_pu}
\alias{lcc_summary_by_pu}
\title{Summarize land cover changes by planning unit}
\usage{
lcc_summary_by_pu(
  crosstab_tbl,
  pu_column,
  pu_name,
  sankey_area_cutoff,
  n_top_lcc = 10
)
}
\arguments{
\item{crosstab_tbl}{A data frame containing at least 2 columns of land cover types (could be character, numeric, or factor) and a "Freq" column containing numeric values.}

\item{pu_column}{A character string representing the column name of the planning unit in \code{crosstab_tbl}.}

\item{pu_name}{A character string representing the name of the planning unit.}

\item{sankey_area_cutoff}{A numeric value indicating the minimum area of changes to be displayed in the Sankey plot.}

\item{n_top_lcc}{An integer representing the number of top land cover changes to be displayed. Default is 10.}
}
\value{
A list containing the Sankey plot (\code{sankey_pu}) and a data frame of the top land cover changes (\code{luc_top_pu}).
}
\description{
This function provides a summary of land cover changes for specified planning units. It returns a sankey plot and the top land cover changes.
}
