% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcc_summary_by_pu.R
\name{lcc_trajectory_by_pu}
\alias{lcc_trajectory_by_pu}
\title{Land Cover Change Trajectory by Planning Unit}
\usage{
lcc_trajectory_by_pu(crosstab_tbl, pu_column, pu_name)
}
\arguments{
\item{crosstab_tbl}{A data frame containing cross-tabulation data.}

\item{pu_column}{A character string specifying the column name of the planning unit in the data frame.}

\item{pu_name}{A character string specifying the name of the planning unit to filter by.}
}
\value{
A list containing the filtered cross-tabulation table and the land cover change trajectory bar plot.
}
\description{
This function filters a cross-tabulation table based on a specified planning unit and generates a bar plot of the land cover change trajectory.
}
