% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_traj_reclass}
\alias{lookup_traj_reclass}
\title{PreQUES Land Use Change Analysis Report}
\format{
A R Markdown (qmd) file that can be rendered into a variety of formats including HTML, PDF, and Word.

A R Markdown (qmd) child file that is incorporated into a main R Markdown file via the \code{child} parameter in a code chunk.

A data frame with 23 rows and 4 columns:
\describe{
\item{Value}{Reclassified land cover codes ranging from 1 to 23}
\item{PL20}{Descriptions for each corresponding land cover code in Bahasa Indonesia}
\item{traj_id}{Trajectory ID for each land cover code}
\item{trajectory}{English descriptions for each corresponding trajectory ID}
}

A data frame with 23 rows and 4 columns:
\describe{
\item{Value}{Reclassified land cover codes ranging from 1 to 23}
\item{PL20}{Descriptions for each corresponding land cover code in Bahasa Indonesia}
\item{traj_id}{Trajectory ID for each land cover code}
\item{trajectory}{English descriptions for each corresponding trajectory ID}
}
}
\source{
The analysis methodology is based on the Quantification of Environmental Services module.
Analysis Per Planning Unit Child Document

This child document is a part of the Quantification of Environmental Services module.
Lookup Table for Reclassified Trajectories

The data was adapted from the Indonesian Ministry of Environment and
Forestry (KLHK).

The data was adapted from the Indonesian Ministry of Environment and
Forestry (KLHK).
}
\usage{
lookup_traj_reclass

lookup_traj_reclass
}
\description{
\code{PreQUES_report.qmd} is a R Markdown document that generates a report for analyzing land use change.
It follows the Quick Ecosystem Service Assessment (Ques) methodology and is used to identify and evaluate the impact of
land use change on various ecosystem services.

\code{quespre_by_pu.qmd} is a knitr child document used in the generation of land use change analysis reports.
This child document is used to perform analyses on a per planning unit basis, thus providing more granular results.

This lookup table maps land cover codes to their respective descriptions and
trajectory classes. The original land cover values have been replaced with a sequence
of numbers from 1 to 23.

This lookup table maps land cover codes to their respective descriptions and
trajectory classes. The original land cover values have been replaced with a sequence
of numbers from 1 to 23.
}
\keyword{datasets}
