% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_trajectory_complete}
\alias{lookup_trajectory_complete}
\title{Lookup Table for Land Cover Change Trajectories}
\format{
A data frame with 64 rows and 7 columns:
\describe{
\item{ID_1}{Numeric identifier for the initial land cover class.}
\item{ID_2}{Numeric identifier for the final land cover class.}
\item{CLASS_1}{Description of the initial land cover class.}
\item{CLASS_2}{Description of the final land cover class.}
\item{ID_traj}{Unique identifier for each land cover change trajectory, formed by
combining ID_1 and ID_2 with an underscore.}
\item{trajectory}{Detailed description of each land cover change trajectory.}
\item{def}{Simplified definition or category of each land cover change trajectory.}
}
}
\source{
LUMENSR package
}
\usage{
lookup_trajectory_complete
}
\description{
The \code{lookup_trajectory_complete} data frame provides a comprehensive lookup table for
land cover change trajectories. It maps the combinations of initial and final land
cover classes to specific trajectory IDs and defines the corresponding trajectory
descriptions.
}
\details{
The \code{lookup_trajectory_complete} data frame provides a mapping of land
cover change trajectories. It considers all possible combinations of initial and final
land cover classes and assigns a unique trajectory ID to each combination. The data
frame also includes detailed descriptions and simplified definitions for each
trajectory.

The land cover classes considered in this lookup table are:
\enumerate{
\item Undisturbed forest
\item Logged-over forest
\item Monoculture tree-based plantation
\item Shrub, grass, and cleared land
\item Agriculture/annual crop
\item Mixed tree-based plantation
\item Others
\item Settlement and built-up area
}

The \code{ID_traj} column is formed by concatenating the \code{ID_1} and \code{ID_2} values with an
underscore, providing a unique identifier for each land cover change trajectory.

The \code{trajectory} column provides a detailed description of each land cover change
trajectory, indicating the specific transition from the initial land cover class to
the final land cover class.

The \code{def} column provides a simplified definition or category for each land cover
change trajectory, such as "Stable forest," "Reforestation," "Deforestation," or
"Other."
}
\examples{
# View the first few rows of the lookup table
head(lookup_trajectory_complete)

# Filter the lookup table for a specific initial land cover class
subset(lookup_trajectory_complete, ID_1 == 1)

}
\keyword{datasets}
