% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUMENSR-package.R
\docType{package}
\name{LUMENSR-package}
\alias{LUMENSR}
\alias{LUMENSR-package}
\title{LUMENSR: An R package for land use planning and management of multiple environmental services}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This proposal suggests an experiment to package the core functions of LUMENS as an R package. By packaging the functions, LUMENS developers could benefit from improved code organization, version control, and testing capabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://icraf-indonesia.github.io/LUMENSR/}
}

}
\author{
\strong{Maintainer}: Dony Indiarto \email{d.indiarto@cifor-icraf.org} (\href{https://orcid.org/0000-0001-9546-8201}{ORCID})

}
\keyword{internal}
