% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUMENSR_example.R
\name{LUMENSR_example}
\alias{LUMENSR_example}
\title{Get path to LUMENSR example}
\usage{
LUMENSR_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
LUMENSR comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
LUMENSR_example()
LUMENSR_example("ntt_temp.tif")
}
