% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntt_admin}
\alias{ntt_admin}
\title{a clipped raster dataset of the BIO1 = Annual Mean Temperature over East Nusa Tenggara
District boundaries of East Nusa Ternggara}
\format{
\subsection{\code{ntt_admin}}{

A polygon 22 features and 2 fields:
\describe{
\item{ID}{Country name}
\item{Kabupaten}{District names of East Nusa Tenggara}
...
}
#' @source \url{https://data.humdata.org/dataset/cod-ab-idn}
}
}
\usage{
ntt_admin
}
\description{
a clipped raster dataset of the BIO1 = Annual Mean Temperature over East Nusa Tenggara
District boundaries of East Nusa Ternggara
}
\keyword{datasets}
