% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntt_admin_lookup}
\alias{ntt_admin_lookup}
\title{District boundaries of East Nusa Tenggara in spatraster format}
\format{
A data frame with 22 features and 2 fields:
\describe{
\item{ID}{Country name}
\item{Kabupaten}{District names of East Nusa Tenggara}
}
}
\usage{
ntt_admin_lookup
}
\description{
A data frame containing the district boundaries of East Nusa Tenggara.
}
\keyword{datasets}
