% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_trajectory.R
\name{plot_bar_trajectory}
\alias{plot_bar_trajectory}
\title{Plot a Bar Chart of Trajectories}
\usage{
plot_bar_trajectory(df)
}
\arguments{
\item{df}{A data frame containing the data to plot. The first column should be the categories and the second column should be the numeric values.}
}
\value{
A ggplot2 object representing the bar plot.
}
\description{
This function creates a bar plot from a data frame with custom colors based on category names.
Categories containing "loss", "recovery", or "stable" are colored salmon, lightgreen, or lightblue, respectively.
All other categories are colored grey.
}
\examples{
\dontrun{
plot_bar_trajectory(table_traj_area)
}
}
