% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical_raster.R
\name{plot_categorical_raster}
\alias{plot_categorical_raster}
\title{Plot categorical raster}
\usage{
plot_categorical_raster(raster_object)
}
\arguments{
\item{raster_object}{A categorical raster file (an object of class SpatRaster)}
}
\value{
A ggplot map
}
\description{
This function plots a categorical raster file using ggplot2.
}
\examples{
\dontrun{
kalbar_LC11 <- LUMENSR_example("kalbar_LC11.tif")
kalbar_LC11 <-   terra::rast(kalbar_LC11)
kalbar_LC11_att <- add_legend_to_categorical_raster(
  raster_file = kalbar_LC11,
  lookup_table = lc_lookup_klhk)
plot_categorical_raster(kalbar_LC11_att)
}
}
