% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lc_freq.R
\name{plot_lc_freq}
\alias{plot_lc_freq}
\title{Land Cover Frequency Plot}
\usage{
plot_lc_freq(lc_table, column_lc_type, column_T1, column_T2)
}
\arguments{
\item{lc_table}{A data frame containing land cover types and their frequency in two different years.}

\item{column_lc_type}{The column name of land cover types in lc_table.}

\item{column_T1}{The column name of the frequency data for Timepoint 1 in lc_table.}

\item{column_T2}{The column name of the frequency data for Timepoint 2 in lc_table.}
}
\value{
A ggplot object representing the side by side bar plot.
}
\description{
This function generates a side by side bar plot comparing the frequency of land cover types in two different years.
}
\examples{
plot_lc_freq(lc_table = lc_tbl,
             column_lc_type = "Jenis tutupan lahan",
             column_T1 = "NTT_LC90_count",
             column_T2 = "NTT_LC20_count")
}
