% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lcc_freq_bar.R
\name{plot_lcc_freq_bar}
\alias{plot_lcc_freq_bar}
\title{Plot Frequency of Land Cover Change as a Bar Chart}
\usage{
plot_lcc_freq_bar(lcc_table, col_T1 = NULL, col_T2 = NULL, Freq = NULL)
}
\arguments{
\item{lcc_table}{A data frame containing at least two columns of characters and one column of numeric values.}

\item{col_T1}{First column of characters. If NULL, the function will assign the first column containing characters.}

\item{col_T2}{Second column of characters. If NULL, the function will assign the second column containing characters.}

\item{Freq}{Numeric column representing frequencies. If NULL, the function will assign the first column containing numeric values.}
}
\value{
A ggplot object showing the frequency of land cover changes as a bar chart.
}
\description{
This function takes in a data frame and plots the frequency of land cover changes as a bar chart.
}
\examples{
plot_lcc_freq_bar(lcc_table = luc_top_10, col_T1 = NULL, col_T2 =NULL, Freq =NULL)
}
