% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sankey.R
\name{prepare_sankey}
\alias{prepare_sankey}
\title{Prepare data for creating a Sankey plot with multiple transitions}
\usage{
prepare_sankey(df, col_order, value_col)
}
\arguments{
\item{df}{The data frame containing the data.}

\item{col_order}{A vector of column names in the order of the transitions.}

\item{value_col}{The name of the column to use as the values of the links.}
}
\value{
A list containing two data frames: 'nodes' and 'links'.
}
\description{
This function prepares a data frame for creating a Sankey plot with multiple
transitions. The Sankey plot represents the flow from one set of values to
another, where the width of the flow is proportional to its quantity.
}
\author{
Dony Indiarto
}
