% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_frequency_table.R
\name{print_frequency_table}
\alias{print_frequency_table}
\title{Prints a frequency table as an HTML table using the gt package}
\usage{
print_frequency_table(
  df,
  title = "Top 10 Type of Land Cover Change",
  labels = NULL,
  freq_col = "Freq",
  table_width = 700,
  font_size = 12
)
}
\arguments{
\item{df}{The data frame to convert into an HTML table.}

\item{title}{The title for the table. Default: "Top 10 Type of Land Cover Change".}

\item{labels}{A named vector of column labels to use in place of the original column names.
Names of the vector should match the column names in the df. Default: NULL (original names are used).}

\item{freq_col}{The name of the column to format as frequency. Default: "Freq".}

\item{table_width}{The width of the table in pixels. Default: 700.}

\item{font_size}{The font size in pixels. Default: 12.}
}
\value{
A gt table object.
}
\description{
Prints a frequency table as an HTML table using the gt package
}
\author{
Dony Indiarto
}
