% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ques_pre_trajectory.R
\name{ques_pre_trajectory}
\alias{ques_pre_trajectory}
\title{Calculate land use and land cover change trajectory}
\usage{
ques_pre_trajectory(
  lc_t1_,
  lc_t2_,
  admin_,
  lookup_traj_reclass,
  lookup_trajectory_complete,
  trajectory_column_name = "trajectory",
  convert_to_Ha = TRUE
)
}
\arguments{
\item{lc_t1_}{A raster layer representing land cover at time 1.}

\item{lc_t2_}{A raster layer representing land cover at time 2.}

\item{admin_}{A raster layer representing administrative zones.}

\item{lookup_traj_reclass}{A lookup table for reclassifying trajectories.}

\item{lookup_trajectory_complete}{A lookup table for completing trajectories.}

\item{trajectory_column_name}{The name of the trajectory column. Default is "trajectory".}

\item{convert_to_Ha}{Logical. If TRUE, the pixel counts are converted to hectares. Default is TRUE.}
}
\value{
A list containing the trajectory map, the area table, and the bar plot.
}
\description{
This function calculates the trajectory map, creates a cross-tabulation of land cover and administrative zones,
and plots a bar chart of the trajectory data.
}
\examples{
\dontrun{
ques_pre_trajectory(lc_t1_, lc_t2_, admin_, lookup_traj_reclass, lookup_trajectory_complete)
}
}
