% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_multipolygon.R
\name{rasterise_multipolygon}
\alias{rasterise_multipolygon}
\title{Rasterize an sf MULTIPOLYGON object}
\usage{
rasterise_multipolygon(sf_object, raster_res = c(100, 100), field = "ID")
}
\arguments{
\item{sf_object}{An sf MULTIPOLYGON object. It must contain an attribute table, with at least one categorical ID (numeric).}

\item{raster_res}{A numeric vector specifying the resolution of the raster. Default is c(100,100).}

\item{field}{A character string specifying the field name to be used for rasterization from the sf object. Default is "ID".}
}
\value{
A SpatRaster object that is a rasterized version of the input sf object, with a legend derived from the attribute table of the sf object.
}
\description{
This function rasterizes an sf MULTIPOLYGON object to a SpatRaster object. The function also retains
an attribute table from the sf object, by assigning categorical ID values to the raster values.
The rasterized SpatRaster object will also contain a legend derived from the attribute table of the sf object.
}
\examples{
rasterise_multipolygon(sf_object = ntt_admin, raster_res = c(100,100), field = "ID")
}
