% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_raster_to_categories.R
\name{reclass_raster_to_categories}
\alias{reclass_raster_to_categories}
\title{Reclassify a Raster to Categories}
\usage{
reclass_raster_to_categories(raster_map, reclass_table)
}
\arguments{
\item{raster_map}{A raster object to be reclassified.}

\item{reclass_table}{A data frame containing the reclassification rules. The first column must be the existing
values (ID_check) and the second column should be the new values (reclass_mat).}
}
\value{
A reclassified factor raster.
}
\description{
This function reclassifies a raster using a provided reclassification table.
}
\examples{
\dontrun{
# assuming we have a raster object "my_raster" and a reclassification table "my_table"
my_reclassed_raster <- reclass_raster_to_categories(my_raster, my_table)
}

}
