% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simple_landcover.R
\name{simple_landcover_plot}
\alias{simple_landcover_plot}
\title{Generate ggplot objects for each raster}
\usage{
simple_landcover_plot(raster_object)
}
\arguments{
\item{raster_object}{The raster object to be plotted.}
}
\value{
A ggplot object.
}
\description{
This function creates ggplot objects to visualize raster data.
}
