% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_map_temperature.R
\name{vis_map_temperature}
\alias{vis_map_temperature}
\title{Visualize Temperature Map}
\usage{
vis_map_temperature(
  temperature_data,
  boundary_data,
  boundary_data_column = "Kabupaten",
  temp_limit = c(10, 35)
)
}
\arguments{
\item{temperature_data}{A spatial raster object that contains temperature data.}

\item{boundary_data}{A spatial polygon object that contains boundary data.}

\item{boundary_data_column}{column names relevant for mapping}

\item{temp_limit}{temperature range in degree celcius, default is c(10,35)}
}
\value{
A ggplot object that displays the temperature map.
}
\description{
This function generates a ggplot object that displays temperature raster data
and boundary polygon data on a map. The color scale of the temperature data is
customized and the legend is labeled in degrees Celsius.
}
\examples{
library(terra)
library(sf)
temp_data <- rast("inst/extdata/east_nusa_tenggara/climate/wc2.1_30s_bio_1.tif")
boundary_data <- st_read("inst/extdata/east_nusa_tenggara/admin/batas_adm.shp")
vis_map_temperature(temperature_data = temp_data, boundary_data = boundary_data)
}
