% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rasters_to_adir.R
\name{write_rasters_to_adir}
\alias{write_rasters_to_adir}
\title{Write SpatRaster Objects to a Directory}
\usage{
write_rasters_to_adir(raster_objects, a_dir)
}
\arguments{
\item{raster_objects}{Named list of SpatRaster objects to write.}

\item{a_dir}{directory for saving the SpatRaster objects.}
}
\description{
This function writes any number of SpatRaster objects to a specified a directory.
}
\examples{
\dontrun{
library(terra)
# assuming rasters is a named list of SpatRaster objects
a_dir<- tempdir()
write_rasters_to_tempdir(rasters, a_dir)
}
}
