% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sankey.R
\name{create_sankey}
\alias{create_sankey}
\title{Create a Sankey Diagram}
\usage{
create_sankey(
  freq_table,
  area_cutoff = 10000,
  change_only = FALSE,
  color_scale = NULL
)
}
\arguments{
\item{freq_table}{A frequency table containing land cover T1, T2, T3, etc.. (character) and a Frequency column (numeric).}

\item{area_cutoff}{Minimum number of pixels of land use land cover frequency to include.}

\item{change_only}{Logical flag, if TRUE exclude persistent land cover.}
}
\value{
A Sankey plot.
}
\description{
This function takes a frequency table (crosstab) as input and creates a Sankey diagram
}
\examples{
library("LUMENSR")
raster_files <- c("kalbar_LC11.tif", "kalbar_LC15.tif", "kalbar_LC20.tif") \%>\%
map(LUMENSR_example) \%>\% map(rast)
raster_list <- map(raster_files, ~apply_lookup_table(raster_file = .x, lookup_lc = lc_lookup_klhk))
crosstab_result <- create_crosstab(raster_list)
create_sankey(freq_table = crosstab_result, area_cutoff = 10000, change_only = FALSE)
}
\author{
Dony Indiarto
}
