% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntt_admin}
\alias{ntt_admin}
\title{Historical Temperature data for East Nusa Tenggara}
\format{
A RasterLayer object with:
\describe{
\item{dimensions}{number of rows, columns, and layers}
\item{resolution}{x and y resolution}
\item{extent}{geographical extent}
\item{crs}{coordinate reference system}
\item{values}{raster values, in this case representing annual mean temperature}
}

A RasterLayer object with:
\describe{
\item{dimensions}{number of rows, columns, and layers}
\item{resolution}{x and y resolution}
\item{extent}{geographical extent}
\item{crs}{coordinate reference system}
\item{values}{raster values, in this case representing annual mean temperature}
}

A SpatialPolygonsDataFrame object with 22 features and 2 fields:
\describe{
\item{ID}{Country name}
\item{Kabupaten}{District names of East Nusa Tenggara}
}
}
\source{
\url{https://www.worldclim.org/data/worldclim21.html}
Historical Temperature data for Indonesia

\url{https://www.worldclim.org/data/worldclim21.html}
District boundaries of East Nusa Tenggara

\url{https://data.humdata.org/dataset/cod-ab-idn}
}
\usage{
ntt_admin
}
\description{
A subset of WorldClim version 2.1 climate data for 1970-2000. This version was released in January 2020; 30 seconds (~1 km2) spatial resolution.
A clipped raster dataset of the BIO1 = Annual Mean Temperature over East Nusa Tenggara.

A subset of WorldClim version 2.1 climate data for 1970-2000. This version was released in January 2020; 30 seconds (~1 km2) spatial resolution.
A clipped raster dataset of the BIO1 = Annual Mean Temperature over Indonesia.

A spatial dataset containing the district boundaries of East Nusa Tenggara.
}
\keyword{datasets}
