% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical_raster.R
\name{plot_categorical_raster}
\alias{plot_categorical_raster}
\title{Plots a categorical raster map}
\usage{
plot_categorical_raster(raster_object)
}
\arguments{
\item{raster_object}{A raster object.}
}
\value{
A ggplot object.
}
\description{
This function takes a raster object as input and produces a ggplot. If the raster
object includes a "color_pallete" column with hex color codes, these colors are
used for the fill scale. Otherwise, the default \code{scale_fill_hypso_d()} fill scale
from the tidyterra package is used.
}
